Adds: process sqs message immediately when..
Adds support for processing the SQS message immediately (rather than
waiting for the configured delay, which is the default) for messages
generated as a direct result of organic updates in the Bib/Item service.
Updates triggered by the Bib/Item service do not require any delay; The
delay is used to accomodate slow data propogation *to* the Bib/Item
service, which is not a concern if the SQS message is created after
those updates have been made.