Adds "de-duping" to resque handler
Adds basic de-duping to resque handler to avoid syncing any barcode that
appears to be redundant. This is done by maintaining a hash inside the
resque handler that associates barcodes with "last sync time". Any time
a new barcode is considered for syncing, the time the barcode was
written to SQS ("queued_at") is examined and compared to the "last sync
time" for that barcode. If it's determined that the same barcode was
actually already synced *after* the "queued_at" time, the barcode is
discarded. This addresses an issue where the resque queue grows so large
that multiple messages for a single barcode may enter the queue and take
so long to be processed that processing the first effectively makes
latter messages redundant (because processing later messages for that
same barcode would fetch the exact same data from the ItemService).

https://jira.nypl.org/browse/SCC-1344