Add default for queued_at
To address a large backlock of redis queue messages that do not have
queued_at property applied, set a conservative default "queued_at" equal
to the date that feature was deployed. Given that greater values of
"queued_at" improve the chance the event will be played (rather than
skipped), selecting the maximum possible time value as a default is a
conservative choice.