Fix: sync-time lookup persisted to redis
This fixes a bug in the "sync-times" lookup, which was recently added
to ensure that barcodes are not synced more often than they reasonably
need to be (by comparing the date the sync request was queued with the
last time the barcode was synced). The lookup was previously stored as a
class variable. Apparently Resque rebuilds the environment *for each
message processed* by design, so class variables are wiped out. This fix
retains most of the original logic but uses Redis to persist the lookup.
Rather than store barcodes in a class variable hash, barcode sync times
are now persisted to redis (e.g. "sync-time-12345" =>
"1556900532788.04"). Redis keys are set to expire in 7 days.