Change: ResqueMessageHandler skips unavailable
This updates ResqueMessageHandler#update to skip generating scsb-xml,
submitting said xml, and calling refile on any SCSB item found to not be
"Available" in SCSB. We know that the update will fail if the item is
anything but Available, so let's skip making expensive http calls. This
implementation differs in effect in another way: by aborting updates for
non-available items, we're not allowing the anticipated error from being
sent to the requesting user (i.e. 'user_email' attached to job). Adds
tests for new availability identification and ResqueMessageHandler#update
generally.