Fix bad unavailable-barcodes computation
Fixes bug where - when determining what barcodes to send to SCSB - the
unavailable subset of barcodes is derived from the already-filtered
available barcodes (preventing any unavailable barcodes from ever being
noted in the log. This fixes that by extracting "unavailable" barcodes
from the original barcode mapping.

This also includes a change to email the requestor ('user_email') when
the barcode to be updated is skipped due to its being not available in
SCSB.